refactor(lib.h): moved task structs to taskqueue.h
