+++ include/hutoken/lib.h
@@ -4,40 +4,6 @@
 #include "Python.h"
 #include <stdbool.h>
 
-struct EncodeContext{
-    bool initialized_encode;
-    struct HashMap* vocab_encode;
-    char* pattern;
-    char* special_chars[256];
-    char* prefix;
-    bool is_byte_encoder;
-};
-
-struct DecodeContext {
-    bool initialized_decode;
-    char** vocab_decode;
-    int vocab_size_decode;
-    char* special_chars[256];
-    char* prefix;
-    bool is_byte_encoder;
-};
-
-struct EncodeTask{
-    char* text;
-    struct EncodeContext* ctx;
-    int* tokens;
-    int* tokens_size;
-    char* error_msg;
-};
-
-struct DecodeTask{
-    int* tokens;
-    int* tokens_size;
-    struct DecodeContext* ctx;
-    char* result;
-    char* error_msg;
-};
-
 PyObject* p_bpe_train(PyObject* self, PyObject* args);
 PyObject* p_bbpe_train(PyObject* self, PyObject* args);
 PyObject* p_encode(PyObject* self, PyObject* args);
